+++ packages/ui-library/src/components/index.tsx
@@ -90,4 +90,4 @@ export * from './customer-satisfaction';
 export * from './thumbs-satisfaction';
 export * from './hidden-number';
 export * from './twitter-share-link-button';
-export * from './news-banner';
+export * from './coin-wrapper';
